base pfs interfaces
