+++ src/pfs/drive/drive.go
@@ -1,4 +1,4 @@
-package driver
+package drive
 
 import (
 	"io"
