+++ src/pfs/serve/api_server.go
@@ -1,4 +1,4 @@
-package server
+package serve
 
 import (
 	"golang.org/x/net/context"
