+++ src/pfs/serve/streaming_bytes_writer.go
@@ -1,4 +1,4 @@
-package server
+package serve
 
 import (
 	"bufio"
